refactor UnaryMinus serde
